int->size_t

+++ src/entities/entity_manager.cpp
@@ -6,13 +6,13 @@ EntityManager::EntityManager() {}
 void EntityManager::addEntity(std::unique_ptr<Entity> e) { entities_to_add.push_back(std::move(e)); }
 
 void EntityManager::processChangedEntities() {
-	for (int i = 0; i < entities_to_add.size(); i++) {
+	for (size_t i = 0; i < entities_to_add.size(); i++) {
 		entities.push_back(std::move(entities_to_add.at(i)));
 		entities.back()->enter();
 	}
 
-	for (int i = 0; i < entities_to_remove.size(); i++) {
-		for (int j = 0; j < entities.size(); j++) {
+	for (size_t i = 0; i < entities_to_remove.size(); i++) {
+		for (size_t j = 0; j < entities.size(); j++) {
 			if (entities_to_remove[i] == entities[j].get()) {
 				entities[j].reset();
 				entities.erase(entities.begin() + j);
@@ -27,7 +27,7 @@ void EntityManager::processChangedEntities() {
 
 void EntityManager::update(double dTime) {
 	processChangedEntities();
-	for (int i = 0; i < entities.size(); i++) {
+	for (size_t i = 0; i < entities.size(); i++) {
 		entities[i]->update(dTime);
 	};
 }
